Refactor Deeplinking implementation

+++ app/controllers/admin/lti/resources.php
@@ -106,7 +106,7 @@ class Admin_Lti_ResourcesController extends AdminBaseController
 
         $deploymentsCount = ResourceLink::countBySql("deployment_id = ?", [$resourceLink->deployment_id]);
 
-        if ($this->launch_type !== 'deep_linking' && $deploymentsCount === 1) {
+        if ($this->launch_type !== 'deep_linking' && $deploymentsCount === 1 && !$resourceLink->deployment->is_default) {
             $resourceLink->deployment->delete();
         }
